# api/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/order/me/pending": {
      "get": {
        "operationId": "OrderController_getUserOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A list with the pending user orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/order/me/{order}": {
      "get": {
        "operationId": "OrderController_getUserOrder",
        "parameters": [
          {
            "name": "order",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The detailled user's order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      },
      "delete": {
        "operationId": "OrderController_cancelOrder",
        "parameters": [
          {
            "name": "order",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/order/from-cart": {
      "post": {
        "operationId": "OrderController_makeOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderFromCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/order": {
      "get": {
        "operationId": "OrderController_getOrders",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The maximum amount of orders to query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "The offset",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "states",
            "required": false,
            "in": "query",
            "description": "The states, comma separated. If empty, all the states will be selected",
            "example": "WAITING_FOR_ACCEPTATION,PREPARATING,DELIVERING",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the selected orders sorted by the createdAt field, with the basic product fields populated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/order/{order}": {
      "get": {
        "operationId": "OrderController_getOrderDetails",
        "parameters": [
          {
            "name": "order",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order details with populated data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/order/{order}/state": {
      "patch": {
        "operationId": "OrderController_changeOrderState",
        "parameters": [
          {
            "name": "order",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrderStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/order/product/{product}": {
      "get": {
        "operationId": "OrderController_getProductOrders",
        "parameters": [
          {
            "name": "product",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the last 20 orders for a certain product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductOrderEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Order"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
...
    "/api/v1/product/{slug}": {
      "get": {
        "operationId": "ProductController_getProduct",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stat",
            "required": false,
            "in": "query",
            "description": "If the view count of the product should be increased",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The details of the products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEntity"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "Product"
        ]
      }
    },
    "/api/v1/product/{id}": {
      "patch": {
        "operationId": "ProductController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Product"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductController_deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The deleted product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Product"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Shop LCJ",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "available": {
            "type": "boolean"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "imagesUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CartItemEntity": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string",
            "description": "The product ID"
          },
          "count": {
            "type": "number",
            "description": "The product count",
            "default": 1
          }
        },
        "required": [
          "product",
          "count"
        ]
      },
      "OrderEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "The order's ID"
          },
          "user": {
            "type": "string",
            "description": "The ID of the user that issued the order"
          },
          "status": {
            "type": "string",
            "description": "The current status of the order",
            "enum": [
              "WAITING_FOR_ACCEPTATION",
              "PREPARATING",
              "DELIVERING",
              "COMPLETED",
              "USER_CANCELLED",
              "ADMIN_CANCELLED"
            ],
            "example": "WAITING_FOR_ACCEPTATION"
          },
          "items": {
            "description": "The list of the products in the order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemEntity"
            }
          },
          "comment": {
            "type": "string",
            "description": "The optional comment the user made on his order"
          },
          "createdAt": {
            "type": "string",
            "description": "The order's creation date",
            "example": "2021-01-01T12:00:00.000Z"
          },
          "modifiedAt": {
            "type": "string",
            "description": "The order's last update date",
            "example": null
          }
        },
        "required": [
          "_id",
          "user",
          "status",
          "items",
          "createdAt"
        ]
      },
      "OrderRecommendationDto": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string",
            "description": "The ID of the product"
          },
          "count": {
            "type": "number",
            "description": "The number of products",
            "minimum": 1,
            "maximum": 10,
            "default": 1
          }
        },
        "required": [
          "product"
        ]
      },
      "OrderFromCartDto": {
        "type": "object",
        "properties": {
          "recommendations": {
            "description": "A list containing the chosen recommended products IDs and their count",
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderRecommendationDto"
            }
          },
          "comment": {
            "type": "string",
            "description": "The optional comment the user made on his order"
          }
        }
      },
      "CancelOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "The reason of the cancellation"
          }
        }
      },
      "ChangeOrderStateDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "description": "The new state",
            "enum": [
              "WAITING_FOR_ACCEPTATION",
              "PREPARATING",
              "DELIVERING",
              "COMPLETED",
              "USER_CANCELLED",
              "ADMIN_CANCELLED"
            ]
          },
          "comment": {
            "type": "string",
            "description": "A small comment about the state change"
          }
        },
        "required": [
          "state"
        ]
      },
      "ProductOrderEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "The order ID",
            "example": "60bbe561064e56d8c95cbe7d"
          },
          "status": {
            "type": "string",
            "description": "The order status",
            "enum": [
              "WAITING_FOR_ACCEPTATION",
              "PREPARATING",
              "DELIVERING",
              "COMPLETED",
              "USER_CANCELLED",
              "ADMIN_CANCELLED"
            ]
          },
          "user": {
            "type": "object",
            "description": "The ID, firstname and lastname of the user"
          },
          "count": {
            "type": "number",
            "description": "The number of units of the product in the order"
          },
          "createdAt": {
            "type": "string",
            "description": "The order creation date",
            "example": "2021-12-24T23:59:59.999Z"
          },
          "modifiedAt": {
            "type": "string",
            "description": "The order last modification date",
            "example": "2021-12-24T23:59:59.999Z"
          }
        },
        "required": [
          "_id",
          "status",
          "user",
          "count",
          "createdAt",
          "modifiedAt"
        ]
      },
...
      "ProductTypeEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "namePluralized": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "name",
          "namePluralized"
        ]
      },
      "CategoryEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "productType": {
            "$ref": "#/components/schemas/ProductTypeEntity"
          }
        },
        "required": [
          "_id",
          "name",
          "slug",
          "productType"
        ]
      },
      "ProductEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "available": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "imagesUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "$ref": "#/components/schemas/CategoryEntity"
          },
          "stockCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "name",
          "slug",
          "coverImageUrl",
          "price",
          "available",
          "description",
          "imagesUrls",
          "category",
          "stockCount"
        ]
      },
      "DeleteProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the product to delete. Used as a confirmation"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}